Add player dialog in run campaign
